pyproject.toml
[project]
name = "ocr-2"
version = "0.1.0"
description = "General-purpose OCR system: PaddleOCR + GPT-4o + LangChain"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "easyocr>=1.7.2",
    "langchain>=1.2.10",
    "langchain-core>=1.2.14",
    "langchain-openai>=1.1.10",
    "langchain-text-splitters>=1.1.1",
    "numpy>=2.2.6",
    "openai>=2.21.0",
    "opencv-python>=4.13.0.92",
    "paddleocr>=3.4.0",
    "paddlepaddle>=3.3.0",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.45.0",
]
